+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.5.0] 2025-04-27
+
+-   Security fixes of indirect deps [PR96](https://github.com/TanguyLe/HoulgateFestBack/pull/96)
+-   Add a route to get the current Paris time [PR97](https://github.com/TanguyLe/HoulgateFestBack/pull/97)
+
 ## [v0.4.1] 2023-06-01
 
 ### Fixed